Issue/73/thesis cleanup, closes #73

+++ research/scripts/multi_sim_script.py
@@ -19,11 +19,18 @@ def check_extra_params(params):
         # params['true_shape'] = int(params['true_shape'][0])
         params['true_loc'] = float(params['true_loc'][0])
         params['true_scale'] = 1./float(params['true_scale'][0])
-
     if 'interim_prior' not in params:
         params['interim_prior'] = 'flat'
     else:
         params['interim_prior'] = str(params['interim_prior'][0])
+    if 'wrong_prior' not in params:
+        params['wrong_prior'] = 0
+    else:
+        params['wrong_prior'] = int(params['wrong_prior'][0])
+        if 'bias_model' not in params:
+            params['bias_model'] = 'flat'
+        else:
+            params['bias_model'] = str(params['bias_model'][0])
     if 'n_galaxies' not in params:
         params['n_galaxies'] = 10**4
     else:
@@ -61,21 +68,6 @@ def make_true(given_key):
         true_scale = test_info['params']['true_scale']
         true_nz = gamma(true_loc, true_scale**2, bounds=(min(test_info['bin_ends']), max(test_info['bin_ends'])))#sps.erlang(true_shape, true_loc, true_scale)
         true_dict = {'amps': [1.], 'means': [true_loc], 'sigmas': [true_scale]}
-            # true_amps = np.array([0.150,0.822,1.837,2.815,3.909,
-            #                       5.116,6.065,6.477,6.834,7.304,
-            #                       7.068,6.771,6.587,6.089,5.165,
-            #                       4.729,4.228,3.664,3.078,2.604,
-            #                       2.130,1.683,1.348,0.977,0.703,
-            #                       0.521,0.339,0.283,0.187,0.141,
-            #                       0.104,0.081,0.055,0.043,0.034])
-            # true_grid = np.linspace(test_info['bin_ends'][0], test_info['bin_ends'][-1], len(true_amps) + 1)
-            # true_grid_mids = (true_grid[1:] + true_grid[:-1]) / 2.
-            # f = spi.interp1d(true_grid_mids, true_amps)
-            # bin_mids = (test_info['bin_ends'][1:] + test_info['bin_ends'][:-1]) / 2.
-            # bin_difs = test_info['bin_ends'][1:] - test_info['bin_ends'][:-1]
-            # true_means = bin_mids
-            # true_amps = f(bin_mids)
-            # true_sigmas = bin_difs
     else:
         bin_range = max(test_info['bin_ends']) - min(test_info['bin_ends'])
         true_amps = np.array([0.20, 0.35, 0.55])
@@ -97,7 +89,7 @@ def make_true(given_key):
 
     return(test_info, true_nz)
 
-def make_interim_prior(given_key):
+def make_interim_prior(given_key, wrong=False, grid=None):
     """
     Function to make the histogram-parametrized interim prior
 
@@ -112,8 +104,17 @@ def make_interim_prior(given_key):
         the discrete distribution that will be the interim prior
     """
     test_info = all_tests[given_key]
+    bin_mids = (test_info['bin_ends'][1:] + test_info['bin_ends'][:-1]) / 2.
+    bin_difs = test_info['bin_ends'][1:] - test_info['bin_ends'][:-1]
 
-    if test_info['params']['interim_prior'] == 'template':
+    if not wrong:
+        int_pr_type = test_info['params']['interim_prior']
+        print('using true interim prior of '+int_pr_type)
+    else:
+        int_pr_type = test_info['params']['bias_model']
+        print('using wrong interim prior of '+int_pr_type)
+
+    if int_pr_type == 'template':
         bin_range = max(test_info['bin_ends']) - min(test_info['bin_ends'])
         int_amps = np.array([0.35, 0.5, 0.15])
         int_means = np.array([0.1, 0.5, 0.9]) * bin_range + min(test_info['bin_ends'])
@@ -124,7 +125,7 @@ def make_interim_prior(given_key):
             int_funcs.append(chippr.gauss(int_means[c], int_sigmas[c]**2))
         interim_prior = chippr.gmix(int_amps, int_funcs,
             limits=(min(test_info['bin_ends']), max(test_info['bin_ends'])))
-    elif test_info['params']['interim_prior'] == 'training':
+    elif int_pr_type == 'training':
         int_amps = np.array([0.150,0.822,1.837,2.815,3.909,
                               5.116,6.065,6.477,6.834,7.304,
                               7.068,6.771,6.587,6.089,5.165,
@@ -132,12 +133,20 @@ def make_interim_prior(given_key):
                               2.130,1.683,1.348,0.977,0.703,
                               0.521,0.339,0.283,0.187,0.141,
                               0.104,0.081,0.055,0.043,0.034])
-        int_grid = np.linspace(test_info['bin_ends'][0], test_info['bin_ends'][-1], len(int_amps) + 1)
-        int_grid_mids = (int_grid[1:] + int_grid[:-1]) / 2.
-        f = spi.interp1d(int_grid_mids, int_amps)
-        bin_mids = (test_info['bin_ends'][1:] + test_info['bin_ends'][:-1]) / 2.
-        bin_difs = test_info['bin_ends'][1:] - test_info['bin_ends'][:-1]
-        int_amps = f(bin_mids)
+        int_grid = np.linspace(0., 1.1, len(int_amps) + 1)
+        int_mids = (int_grid[1:] + int_grid[:-1]) / 2.
+        int_amps = spi.griddata(int_mids, int_amps, bin_mids, method='linear', fill_value=np.min(int_amps), rescale=False)
+        # # int_grid = np.concatenate((int_grid, np.array([test_info['bin_ends'][-1]])))
+        # # print(int_grid)
+        # int_amps.append(int_amps[-1])
+        # int_amps = np.array(int_amps)
+        # # this basically hardcodes in that grids start at 0.!
+        #
+        # int_grid_mids = (int_grid[1:] + int_grid[:-1]) / 2.
+        #
+        # int_grid_mids = np.concatenate((int_grid_mids, np.array([test_info['bin_ends'][-1]])))
+        # f = spi.interp1d(int_grid_mids, int_amps)
+        # int_amps = f(bin_mids)
         int_means = bin_mids
         int_sigmas = bin_difs
         n_mix_comps = len(int_amps)
@@ -167,9 +176,12 @@ def make_catalog(given_key):
 
     test_dir = os.path.join(result_dir, test_name)
     test_info['dir'] = test_dir
-    if os.path.exists(test_dir):
-        shutil.rmtree(test_dir)
-    os.makedirs(test_dir)
+    if test_dir != '':
+        if os.path.exists(test_dir):
+            shutil.rmtree(test_dir)
+        os.makedirs(test_dir)
+    else:
+        print('Check the config file for newlines!')
 
     param_file_name = test_name + '.txt'
     params = chippr.utils.ingest(param_file_name)
@@ -195,8 +207,18 @@ def make_catalog(given_key):
     interim_prior = make_interim_prior(given_key)
 
     posteriors = chippr.catalog(param_file_name, loc=test_dir, prepend=test_name)
-    output = posteriors.create(true_nz, interim_prior, N=test_info['params']['n_gals'])
+    output = posteriors.create(true_nz, interim_prior, N=test_info['params']['n_gals'], vb=True)
+    print('using implicit prior '+str(posteriors.cat['log_interim_prior']))
     # data = np.exp(output['log_interim_posteriors'])
+    # print('bin ends from simulation: '+str(posteriors.bin_ends))
+
+    if test_info['params']['wrong_prior']:
+        int_pr = make_interim_prior(given_key, wrong=True)
+        int_pr_fine = np.array([int_pr.pdf(posteriors.z_fine)])
+        int_pr_coarse = posteriors.coarsify(int_pr_fine)
+        posteriors.cat['log_interim_prior'] = u.safe_log(int_pr_coarse[0])
+    print('writing implicit prior '+str(posteriors.cat['log_interim_prior']))
+
     posteriors.write()
     data_dir = posteriors.data_dir
     with open(os.path.join(data_dir, 'true_params.p'), 'w') as true_file:
